requirements.txt
# List of python packages required by the project
pycryptodome
flask
argon2-cffi
requests
pyotp
qrcode